+++ apps/api/src/mail/emails/signin-code-email.tsx
@@ -5,21 +5,26 @@ import BaseEmailTemplate from './components/base-email-template'
 interface SignInCodeEmailTemplateProps {
   code: string
   name?: string
+  device: string
+  location: string
 }
 
 export const SignInCodeEmailTemplate = ({
   code,
-  name
+  name,
+  device,
+  location
 }: SignInCodeEmailTemplateProps) => {
   return (
     <BaseEmailTemplate
       previewText="Your Sign in Code for Keyshade CLI"
       heading="Keyshade Sign in Code"
     >
-      <Text style={text}>Dear {name ?? 'User'},</Text>
+      <Text style={text}>Hey {name ?? 'there'}!,</Text>
       <Text style={text}>
-        We’ve sent you this email to verify your Keyshade account. Your Sign in
-        Code is:
+        We’ve sent you this email to verify your Keyshade account from{' '}
+        <strong>{location}</strong> using <strong>{device}</strong>. Your Sign
+        in Code is:
       </Text>
       <Section style={workspaceDetails}>
         <Text style={otpStyle}>