+++ apps/api/src/mail/services/interface.service.ts
@@ -5,7 +5,13 @@ export interface IMailService {
 
   sendEmailChangedOtp(email: string, otp: string): Promise<void>
 
-  sendSignInCode(email: string, code: string, name: string): Promise<void>
+  sendSignInCode(
+    email: string,
+    code: string,
+    name: string,
+    device: string,
+    location: string
+  ): Promise<void>
 
   invitedToWorkspace(
     email: string,