+++ apps/cli/src/commands/profile.command.ts
@@ -1,9 +1,8 @@
 import BaseCommand from '@/commands/base.command'
-import CreateProfile from '@/commands/profile/create.profile'
-import DeleteProfile from '@/commands/profile/delete.profile'
 import ListProfile from '@/commands/profile/list.profile'
 import UpdateProfile from '@/commands/profile/update.profile'
-import UseProfile from '@/commands/profile/use.profile'
+import SwitchProfile from '@/commands/profile/switch.profile'
+import RemoveProfile from '@/commands/profile/remove.profile'
 
 export default class ProfileCommand extends BaseCommand {
   getName(): string {
@@ -16,11 +15,10 @@ export default class ProfileCommand extends BaseCommand {
 
   getSubCommands(): BaseCommand[] {
     return [
-      new CreateProfile(),
-      new UpdateProfile(),
       new ListProfile(),
-      new UseProfile(),
-      new DeleteProfile()
+      new SwitchProfile(),
+      new UpdateProfile(),
+      new RemoveProfile()
     ]
   }
 }