+++ packages/schema/src/auth/index.ts
@@ -2,6 +2,16 @@ import { UserSchema } from '@/user'
 import { z } from 'zod'
 import { WorkspaceSchema } from '@/workspace'
 
+export const DeviceDetailSchema = z.object({
+  ipAddress: z.string().ip().optional(),
+  encryptedIpAddress: z.string(),
+  os: z.string().optional(),
+  agent: z.string().optional(),
+  city: z.string().optional(),
+  country: z.string().optional(),
+  region: z.string().optional()
+})
+
 export const ResendOTPRequestSchema = z.object({
   userEmail: z.string().email()
 })
@@ -11,7 +21,9 @@ export const ResendOTPResponseSchema = z.void()
 export const ValidateOTPRequestSchema = z.object({
   email: z.string().email(),
   otp: z.string().length(6),
-  mode: z.enum(['cli']).optional()
+  mode: z.enum(['cli']).optional(),
+  os: z.string().optional(),
+  agent: z.string().optional()
 })
 
 export const ValidateOTPResponseSchema = UserSchema.extend({
@@ -22,7 +34,9 @@ export const ValidateOTPResponseSchema = UserSchema.extend({
 
 export const SendOTPRequestSchema = z.object({
   email: z.string().email(),
-  mode: z.enum(['cli']).optional()
+  mode: z.enum(['cli']).optional(),
+  os: z.string().optional(),
+  agent: z.string().optional()
 })
 
 export const SendOTPResponseSchema = z.void()